refactor: added directories for Spring Boot

+++ lab2/services/src/main/java/Lab3/Services/CatService.java
@@ -1,9 +1,12 @@
-import Entities.Cat;
-import Entities.Owner;
+package Lab3.Services;
+
+import Lab3.Entities.Cat;
+import Lab3.Repositories.CatRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CatService {
@@ -24,13 +27,15 @@ public class CatService {
         return catRepository.findAll();
     }
 
-    public void saveCat(Cat cat) {
-
+    public void addCat(Cat cat) {
+        Cat newCat = new Cat(cat.getName(),cat.getDateOfBirth(),cat.getBreed(),cat.getColor());
         catRepository.save(cat);
     }
 
-    public void deleteCat(Cat cat) {
-
+    public void deleteCat(Long id) {
+        Optional<Cat> deletingCat = catRepository.findById(id);
+        if (deletingCat.isEmpty()) return;
+        Cat cat = deletingCat.get();
         catRepository.delete(cat);
     }
